Blackbox getVirtualServoAngle() fix

+++ src/main/blackbox/blackbox.c
@@ -360,7 +360,7 @@ static bool blackboxModeActivationConditionPresent = false;
 static bool blackboxIsOnlyLoggingIntraframes() {
     return masterConfig.blackbox_rate_num == 1 && masterConfig.blackbox_rate_denom == 32;
 }
-extern float getVirtualServoAngle();
+extern int16_t getVirtualServoAngle();
 
 static bool testBlackboxConditionUncached(FlightLogFieldCondition condition)
 {
